+++ Graph.h
@@ -10,7 +10,7 @@ private:
 	size_t _vexNum;    // 图的顶点个数 number of vertices  
 	size_t _edgeNum;   // 图的边数 number of edges
 	size_t _graphType;   // 图的类型，1代表有向图，2代表无向图, types of graph, directedgraph is 1, undirectedgraph is 2.
-	std::vector<std::vector<double>> _adjMatrix;  //  邻接矩阵, 记录图中顶点和边的关系, adjacency matrix
+	std::vector<std::vector<float>> _adjMatrix;  //  邻接矩阵, 记录图中顶点和边的关系, adjacency matrix
 
 public:
 	//构造函数
@@ -23,7 +23,7 @@ public:
 	bool checkVexnum(size_t vexnum, size_t edge);
 
 	// 判断每次输入的的边的信息是否合法
-	bool checkEdgeValue(size_t start, size_t end, double weight);
+	bool checkEdgeValue(size_t start, size_t end, float weight);
 
 	// 返回顶点个数
 	size_t vexNum() const { return _vexNum; };
@@ -32,7 +32,7 @@ public:
 	size_t edgeNum() const { return _edgeNum; };
 
 	// 返回邻接矩阵
-	std::vector<std::vector<double>> adjMatrix() const { return _adjMatrix; };
+	std::vector<std::vector<float>> adjMatrix() const { return _adjMatrix; };
 
 	//打印邻接矩阵
 	void print();
